Delete the passport from the returned payload.
This is done when the passport is passed "in-band" and it is done
because the targeting edge handler was initially made to return
targeting data directly in the form of 'key values' on the
returned JSON payload -- if we don't delete the `passport`
attribute here, it may end up sent as targeting data to ad
servers. Not the end of the world, but something we want to avoid
due to passport size.