Add support for in-band passport transport (#26)
* Regen targeting/prebid-us-east-16.html

* Add support for in-band passport transport

When configuring an instance of OptableSDK, we now support a `cookies`
attribute (`true` by default) which, when set to `false`, will cause the
SDK to use "in-band" passport transport, more specifically:

- The `passport` parameter will be set in the query string in each
  network call to the sandbox
- The sandbox will return a `passport` attribute in any successful JSON
  response payload
- The SDK will attempt to read the returned `passport` attribute and
  store it in LocalStorage, scoped to the sandbox host and site slug
  with which the OptableSDK instance was instantiated
- At the subsequent network call to the sandbox, any passport value
  previously stored in LocalStorage will be retrieved and passed back to
  the sandbox as the `passport` query string parameter value

Also updates README for in-band passport transport, and adds demos for
the in-band passport transport case which can be browsed from
.../demos/index-nocookies.html

Depends on optable-sandbox PR419:
https://github.com/Optable/optable-sandbox/pull/419

* Delete the passport from the returned payload.

This is done when the passport is passed "in-band" and it is done
because the targeting edge handler was initially made to return
targeting data directly in the form of 'key values' on the
returned JSON payload -- if we don't delete the `passport`
attribute here, it may end up sent as targeting data to ad
servers. Not the end of the world, but something we want to avoid
due to passport size.